Update readme to use recipes/debian.yml consistently
The file `recipes/debian.yml` was added back in 2020 in commit
ab8f0796df8b7d71d60a7503a5cd1e4c8424cb90. The "official way" to enable
Salsa CI pipelines is to include this file, as documented in the TLDR.

Yet the readme was never updated to reflect that. Fixed with this
commit.

This was also discussed about 2 years ago in:
https://salsa.debian.org/salsa-ci-team/pipeline/-/merge_requests/327