Move extract-source tasks to the build job and use sbuild for building
These changes move the extract-source-related tasks to the build job.
For that, the build job needs to build the package in a clean
environment, for the target distribution and architecture, triggered
from the same build job where the debianized source tree was unpacked.
The build job now uses then the build:latest image, based on debian sid,
and it relies now on similar modern tools than those currently used in
the buildd network: sbuild with the unshare backend. The build job also
creates the chroot used to build the package with mmdebstrap.

Since the pipeline no longer needs the artifacts from a previous stage
to build the package, we can expect that more packages will be able to
use the pipeline under the current max artifacts size limitations.
(See #195).

This also completely solves the pretty expensive source job issue
(#187).

It is worth noting that these changes would also make it easier to use
debrebuild, instead of reprotest to check the package reproducibility
(See #399).

Changes include some improvements related to ccache. For example,
instead of zeroing the ccache stats before building, the pipeline shows
the stats to make it easier to compare them with the ccache's cache
status at after the build process.

Closes: #187,#195,#296