Emit clear error messages if git commits are not signed
The single most common error in CI for new contributors currently is
that the commit is not signed, or signed with a key that can't be
fetched from the OpenPGP or Ubuntu keyservers.

Make the error messages more clear. In particular if a commit was not
signed the CI could just exit - now it will exit with error message.

Use YAML scalar `- >` to allow multilines without extra \ syntax and the
end of each line. Use `git log --pretty="%GP"` to avoid complex `awk` and
to simplify code and to have a more robust empty variable check instead
of relying on exit code inside a pipe.

Also fix indentation of git command to be easier to read on a single
line.