Support branch/tags as target heads in test repos
The pipeline's own CI relies on running the pipeline against different
projects that are defined in the test-pipeline section from
.gitlab-ci.yml, that triggers child pipelines for those projects. For
reasons such as: stability on the tests, pipeline fork's tests
specificities, testing against a specific release, it may be needed to
run the pipeline on a specific HEAD of such projects.

This commit makes it possible to define a branch or tag in the projects
to be tested. The target HEAD must be defined in the PROJECT_TARGET_HEAD
variable in the test-pipeline:parallel:matrix section. For example:

test-pipeline:
  trigger:
    include:
      - local: .pipeline-test.yml
    strategy: depend
  ...
  parallel:
    matrix:
      - PROJECT_URL: https://salsa.debian.org/debian/grep
        PROJECT_TARGET_HEAD: debian/3.3-1
        ...