Build autopkgtest image on top of stable
We aim to mimic ci.debian.net as much as possible, so users can expect
the most similar results in both CI "infrastructures". Let's base the
autopkgtest image on top of stable, since it is the case on ci.d.net:
https://lists.debian.org/debian-ci/2024/05/msg00002.html

Use the AUTOPKGTEST_IMAGE_VERSION variable to define the autopkgtest
image tag.

Tag as autopkgtest:latest the same stable-based autopkgtest image.

Closes: #350