Add SALSA_CI_DISABLE_EXPORTORIG_FALLBACK option
In this commit, allow to disable gbp export-orig fallback.
Even though maintainer forgot to push pristine-tar branch,
CI continues to process with origtargz fallback.

It may be harmful because it can be overlooked with this behavior.

By default, SALSA_CI_DISABLE_EXPORTORIG_FALLBACK=0 as same as before
to keep compatibility.