autopkgtest image: do not install tools with -t unstable
Installing tools with `apt-get -t unstable` overrides setting "stable" as
the default release in /etc/apt/apt.conf.d/default-release.conf, and
causes all the dependencies of the packages to install to be resolved in
unstable.

We want to always try to pull as many packages as possible from the
default release (stable), so let's drop the -t setting. We rely on
pinning (/etc/apt/preferences.d/apt-pinning.pref) to pull autopkgtest
and companion tools from unstable.