Fix autopkgtest job bug in internal pipeline (Failure to pull the job's image)
Before this commit the autopkgtest job fail to pull it's docker image
when running in the internal pipeline in default branch.

The reason for that is that autopkgtest job has a different version
variable for it's image and it is not assigned a value in the condition
of running in default branch.

Failed job logs: https://salsa.debian.org/salsa-ci-team/pipeline/-/jobs/5825361

Error Message:
```
ERROR: Job failed: failed to pull image "registry.salsa.debian.org/salsa-ci-team/pipeline/autopkgtest:" with specified policies [always]: invalid reference format (manager.go:250:0s)
```

This commit fixes that by assigning it a value in this condition.