Add a test pipeline to test ARM build jobs
This creates build jobs to test the ARM base images built by the
triggering pipeline. This can be extended for architectures to be
supported in the future (like RISC-V). The test pipeline run by default
on http://salsa.debian.org/salsa-ci-team/pipeline, and it has to be
manually triggered otherwise, including in merge request pipelines