Install sid version of debootstrap at autopkgtest image
To support Ubuntu releases which aren't supported in the debootstrap stable
version (like noble on bookworm). It is safe to remove this apt pinning if Ubuntu
support got deprecated unless other supported distros/forks became dependent on it later.

Example which shows that Ubuntu noble is supported in the sid version but not on
bookworm version (search for noble on those pages):
- Supported releases at current sid version:
  - https://salsa.debian.org/installer-team/debootstrap/-/tree/1.0.137/scripts

- Supported releases at bookworm version:
  - https://salsa.debian.org/installer-team/debootstrap/-/tree/bookworm/scripts