Suppress the bloat log of cache removing
We have encountered many instances where the kernel build log is filled
with messages about deleting the cache files like below:
  removed '/builds/xxx/kernel/linux/../.ccache/7/d/e1dt5tjd3134nhvhrq7upceer4ujnj8M'
  removed '/builds/xxx/kernel/linux/../.ccache/7/d/5cl7mvmrr6hrqpgoo8ii3382v3qlgqqM'
  removed '/builds/xxx/kernel/linux/../.ccache/7/d/cc2tt5d9luo866k6dn7l576riio5lq4R'
  ...

These messages exceed the maximum size of the pipeline job log, making
it impossible for us to see the truly useful kernel build log. This
situation causes significant inconvenience for the kernel debugging.
It really makes no sense to display all these messages in the build log.
Therefore, we choose to limit the output to the last 10 lines of these
messages, allowing us to see if ccache is caching anything without
overwhelming the build log.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>