Disable build_path variation by default in reprotest job
Quoting Holger's comments about why the build_path should be disabled by
default:

The build path is recorded in the .buildinfo file and can be reused when
rebuilding a package.

Reproducibility while varying the build path has no advantages in the
real world (as oppposed to the old CI tests on tests.r-b.o) because
there's no benefit from being able to vary the build path, while on the
other hand there are disadvantages: hundreds of packages become
unreproducible, but just because we artificially make them so. If we
(re)use the known build path, they are/become reproducible. And this is
what we want and thus we don't vary the build path any more.

And that's why the salsa CI pipeline also should not vary the build path
by default.

Closes: #368