Add a way to specify artifact_pattern argument for reprotest
Images that the debian-installer source package creates are put into a
"debian-installer-image_$VERSION_$ARCH.tar.gz" file instead of a binary
package. Furthermore, installer-related packages can generate `*.udeb`
files instead of (or in addition to) `*.deb` ones.

The default reprotest job doesn't diff these files. We cannot use
SALSA_CI_REPROTEST_ARGS to tell reprotest to do so, because it will be
interpreted as the build command argument to reprotest.

Define a SALSA_CI_REPROTEST_ARTIFACT_PATTERN variable that can be used
to tell reprotest which files it needs to consider build artifacts. The
command expects one space-separated argument like "*.dsc *.changes", so
make the variable a single string and quote it for the command.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>