Recommend using a salsa-ci.yml file to activate Salsa CI
Instead of pointing the CI/CD configuration to
`recipes/debian.yml@salsa-ci-team/pipeline`, recommend using a real
`salsa-ci.yml` file on the Debian git branches of the project.

This makes use of Salsa CI more explicit. It is guaranteed to run only
on the Debian packaging branches. Any future customizations to the file
will automatically take effect, and interaction with forks and Merge
Requests will work better as the direct use of `salsa-ci.yml` is explicit.