WIP: Add new documentation about what jobs run on what releases
So far Salsa CI has been focused on testing packages prior to upload to
Debian unstable (sid), while testing packages prior to upload to other
releases has been a secondary feature. The purpose of this documentation
is to establish what jobs run, why and where so we have a shared
understanding of how the pipeline is supposed to morph based on the
RELEASE value. With this documentation established, it is easier to
decide and implement each individual job and its behavior in different
Debian releases, and even other distros (currently Kali and Ubuntu).