Revert "Use lintian:latest image instead of :${RELEASE} in lintian job"
This reverts commit cf2aa40600517feedde0014a65b681d0199606e7 and
https://salsa.debian.org/salsa-ci-team/pipeline/-/merge_requests/347

Lintian should the same version as the target RELEASE has. For updates
of packages in older releases it should be enough that those packages
don't regress for the existing Lintian version. In most cases it would
be directly wrong to do any larger changes for old packages to meet the
requirements of the latest Lintian version.

New versions of Lintian may have newer and stricter requirements
intended for new versions of packages to follow. If there are serious
issues in old Lintian versions, those will be backported to old
releases.