Fix missing double quotes in conditional expression
To be honest, it's not really a fix, bash/sh is happy all the same if
there's no argument for `-z`. For example, the following expression is
valid and doesn't return any error:

    [ -z ] && echo empty
    empty

This being said, `git grep -E -- '(test|\[) -[nz]'` shows that we always
use double quotes for those conditional expressions, except for *this
one here*, so for the sake consistency, let's use double quotes here as
well.

This commit also prefixes the code block with a comment, to follow the
style used in this particular job.