Add support for Kali base docker images
Hence Kali doesn't rely on Debian's testing images anymore, so we can
drop all the code that was used to crossgrade a Debian system into a
Kali system, in the file `images/files/base/initialize`.

For documentation purpose (to help other distros who might need it),
the former code is kept in a comment.