Rename base image argument ARCH -> REGISTRY_NAMESPACE
This is in preparation for the next commit to come.

In fact, the fact that `REGISTRY_NAMESPACE` == `ARCH` isn't true for
every image on the Docker Hub registry. It's the case only for
"official" Docker images (ie. images published by Docker Inc.), such as
Debian or Ubuntu. Those official images are the only ones that are
allowed to have the naming scheme `docker.io/<ARCH>/<IMAGE>`, for
example `docker.io/amd64/debian:unstable`.

Images that are maintained by users of the Docker Hub (such as Kali
Linux) have a different naming scheme: `docker.io/<USERNAME>/<IMAGE>`,
for example `docker.io/kalilinux/kali-dev:amd64`.