Factorize some yaml for self checks
With this commit, there's absolutely no change for the two jobs `check
license + contributor` and `check gpg signature`: we just factorize the
common code in a new hidden job `.self-checks`.

There's a slight change for `check yaml lint`: it didn't have any rule,
and it now inherits the same rules as the two other checks, which I
think is a good thing, for simplicity and consistency.