Use 'after_script' section to run commands needed also on failing builds
If a the 'script' section runs a command that returns a non-zero exit code
(a failure), it will stop executing. Commands that are needed to give
debug information or save cache for the next run should be instead in the
'after_script' section that will run even on failed builds.

This helps in particular very large builds, that may time out, and need
to have their ccache uploaded and used in the next run. The post-build
diagnostic commands are also needed for information about artifacts,
which get uploaded on failing builds as well.