Save cache for next run even on failed builds
The purpose of a cache (and in particular ccache) is to avoid wasting
CPU cycles on re-building the same code over and over. It is useful not
just on successful builds, but also when builds fail. In particular if
building something big and when using SALSA_CI_BUILD_TIMEOUT_ARGS exactly
as instructed in the README, the cache must be uploaded on a failed build
to avoid starting the re-run from zero again.

If is very unlikely that the cache itself caused a build failure. If it
would ever happen, users can always clear caches manually from the GitLab
CI page and restart a job.

Closes: #404