Extend Salsa CI self-tests with galera-4 pipeline to catch more regressions
The source package 'galera-4' is maintained on stable and oldstable, and
it is relatively small to build fast, yet not too trivial to lack features.
Hence it is a good candidate to be used for regression testing of Salsa CI
pipeline changes.

For builds and tests to be able to run on stable, oldstable and
experimental, build the 'base' image on all of them. Additionally build
the additional required images for tests on stable to be able to run.

Also introduce variable $OLDSTABLE so future maintenance is easier.