Suppress the unreleased-changes tag in one of the grep test pipelines
The grep test pipeline that helps to run all the test jobs is run for
the latest grep git commit, that may be UNRELEASED. lintian may comply
about this, throwing an unreleased-changes error. We could run the
pipeline against a tag too, but that means on an actual package release,
and sometimes there are useful or meaningful changes that are
unreleased, e.g. wrap-and-sort.