Disable SALSA_CI_DISABLE_CROSSBUILD_ARM64 to avoid frequent errors
As the cross-build requirement in Debian isn't universal at the moment,
disable this test by default to avoid unnecessary noise. Having a CI
that is consistently green, and that the "greenness" is respected, is
of higher importance than having lots of things tested.

Additionally, don't allow failures on this job anymore. After this
change, anyone having this job enabled by intentionally will need to fix
their package (if possible) or decide to allow failures in their own
pipeline.

While it would be nice to have all packages cross-build, some ecosystems
are inherently not cross-buildable, and many packages are not
cross-buildable while updates in the toolchain are ongoing and there is
version skew between platforms. See https://crossqa.debian.net/ for
latest details.

Having this test enabled by default, and failing on numerous pipelines
led to an influx of bug reports to the Salsa CI pipeline, and created
confusion among packagers about what the problem is and if it is
serious.

Also, this test really tests the cross-build toolchain more than
anything else, so packages with a failing CI could not rarely fix
anything in the package itself.

Packagers can still easily test this with by triggering a manual build
and setting the variable in the GitLab CI pipeline to '1' in an
interface that also includes a one-line description of what the variable
does.

Closes: #374