Replace the build twice job with validate package cleanup
Instead of building a package twice in a row, make it possible to add
opt-in configuration to sbuild to validate if the clean target of
debian/rules correctly restores the source tree to its initial state.
This is handled now with the new SALSA_CI_*_VALIDATE_PACKAGE_CLEAN_UP
variables.