Check that ccache is installed before using it in build-after-script
ccache is installed during `.build-script`. If ever the job fails before
that, or during `.build-before-script`, then ccache is not installed and
then the `after-script` section fails:

```
Running after script...
$ CCACHE_DIR=${CCACHE_TMP_DIR} ccache -s
/bin/bash: line 239: ccache: command not found
WARNING: after_script failed, but job will continue unaffected: exit code 1
```

This is a regression that was introduced in
66fb5b9bc368099fd3d6db0448d182061158d9de.

Note that in the future, we might use the `build` image, introduced in
e123d4257088c96b9475009f0f6167c3d0dffa39, and given that ccache is
installed unconditionnaly in this image, there will be no need for this
check anymore.