Drop link to #salsaci and emphasize Issues and MRs over IRC discussions
The webchat.oftc.net no longer works as new nicks joining the channel need
to be identified and verified first. The link should be removed to not
waste people's time trying to use it, and to avoid situations where
newcomers get put off by broken links and broken communication channels.

Extend the contact and contribute sections with encouragement for anyone
to participate in Issues and Merge Requests as currently commenting them
is the main communication channel for the team.