Add support for Ubuntu 25.10 (Questing Quokka) and use it by default
Ubuntu 25.10 development has started. Add it to the list of
supported Ubuntu releases and also start using it as the default
Ubuntu release as well as the release used for testing regressions
in the Salsa CI pipeline definition itself.

For the pipeline to pass, the salsa-ci-team/autopkgtest-lxc!32
needs to be merged first and autopkgtest image be available.