Fix autopkgtest images cleanup: use STABLE instead of RELEASE
The staging autopkgtest images were never cleaned because they were tagged with $STABLE
release (e.g. bookworm) during build, but attempted to clean with $RELEASE
variable (e.g. sid) in the clean images job resulting in a mismatch and leaving
the images in the registry.

Note: autopkgtest images are only built for $STABLE releases, to match the behavior on ci.debian.net.