Replace timeout command with job variable RUNNER_SCRIPT_TIMEOUT
GitLab 16.4 introduced the RUNNER_SCRIPT_TIMEOUT and RUNNER_AFTER_SCRIPT_TIMEOUT
variables, which are designed for the exact use-case previously handled by the
timeout command. This commit replaces the timeout command with these variables
without altering job behavior.

Using these job variables is preferred as they are more reliable and maintainable.
The previous timeout command did not account for commands run before it and added
an extra layer of complexity.