Use filter-output for blhc job and output artifacts for piuparts
The output for blhc was reported to exceed the job's log limit for some packages.
Using filter-output and providing the logs as artifacts should resolve this issue.
Additionally, the piuparts job already used filter-output but did not output the logs as artifacts.
We now provide the logs as artifacts for piuparts as well, since using filter-output assumes the log limit might be exceeded.

In reprotest job, reprotest.log was used as the log name, so we followed the same convention with blhc and piuparts here.

Closes #225, #299