Make autopkgtest for i386 opt-in
Switch the i386 autopkgtest job to opt-in instead of running by default.
Running i386 tests is only useful for a small number of packages, while
for most it just adds load and confusion.

Since i386 is now mostly used in limited contexts and only a small
number of packages actually need testing on it, disabling it by default
reduces resource usage and pipeline time.

Closes: #484