Fix how ARM jobs are prevented from running by default
Move the `when:` to the top-level YAML so it overrides rules, and
instead of disabling the ARM jobs completely, allow them to still run if
manually triggered. This will make it easier to validate that runners
are properly back online in the near future.

This fixes commit 172c722 where the `when: never` did not have the
intended effect as it was being overridden by the `when: always` in the
rules about $CI_PROJECT_PATH and $CI_DEFAULT_BRANCH.