Clean up unnecessary space and semicolons in if clauses
If a shell `if` statement has the `then` on a new line, using a semicolon
`;` is unnecessary. Also having space in `] ;` is unnecessary, so clean
these away.

For now, don't touch use of `]; then` which could be made easier to read
by having the `then` keyword on a separate line in the future and the
semicolon removed.