Clean away dysfunctional ccache
The ccache functionality has been broken since the pipeline was
refactored to use sbuild. Worse than just not working, it is eating up
disk space making builds of large packages fail in ways that are very
un-intuitive for anyone to debug and recover from.

Remove the ccache functionality so that it is easier for somebody to
re-implement it from scratch in the future. It will be easier to
implement without any baggage from previous version, as trying to
understand the old version is more effort than developing a new version,
and trying to create a new version and maintain artificial backwards
compatibility with something that was broken and didn't work to begin
with is just waste of effort.

After this the floor is clear for anyone to implement a new caching
mechanism that does not even need to be tied to just ccache but can use
any other or even multiple mechanisms to minimize rebuilding the same
code over and over.

Closes: #492