Remove obsolete compatibility code that is now a confusing no-op
`SALSA_CI_APTLY_GPG_PASSPHRASE` used to be misspelled as
`SALSA_CI_APTLY_GPG_PASSPHASE` (notice the missing `R`).  An attempt to
fix this typo while reserving compatibility lead to !194.

Forward 4 years, and there's apparently corner cases where the
compatibility does not work as intended and breaks stuff, leading to
!445 fixing the typo in the compatibility code.

This effectively renders the compatibility code moot, as it falls back
on itself; but it also is very confusing when seeing the construct and
the comment about compatibility.

So, given it's been 2 years since this hasn't been changed, and is most
probably not an issue anymore, remove the compatibility altogether to
simplify things and avoid confusion.